% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cancelled}
\alias{cancelled}
\title{General functions responsible for the different GTM API calls}
\usage{
cancelled()
}
\description{
General functions responsible for the different GTM API calls
}
\seealso{
Other tag functions: \code{\link{gtm_tags_list}}
}
\concept{tag functions}
