% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_accounts}
\alias{get_accounts}
\title{Gets a GTM Account.}
\usage{
get_accounts(accountId)
}
\arguments{
\item{accountId}{The GTM Account ID}
}
\description{
Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/tagmanager.edit.containers, https://www.googleapis.com/auth/tagmanager.manage.accounts, https://www.googleapis.com/auth/tagmanager.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
