% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{get_container_versions}
\alias{get_container_versions}
\title{Get a GTM account's Container version detail}
\usage{
get_container_versions(accountId, containerId, headers, includeDeleted)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{GTM containerId}

\item{headers}{Retrieve headers only when true. (Default: false)}

\item{includeDeleted}{Also retrieve deleted (archived) versions when true. (Default: false)}
}
\value{
container info
}
\description{
Get a GTM account's Container version detail
}
\seealso{
Other get functions: \code{\link{get_accounts_list}},
  \code{\link{get_container}}, \code{\link{get_tags_list}},
  \code{\link{get_triggers_list}},
  \code{\link{get_variables_list}}
}
