% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{get_containers_list}
\alias{get_containers_list}
\title{Get containers list}
\usage{
get_containers_list(accountId)
}
\value{
a list of containers
}
\description{
Get containers list
}
\details{

}
\seealso{
Other list functions: \code{\link{get_enviroments_list}}
}
