% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{get_trigger}
\alias{get_trigger}
\title{Get a trigger's details}
\usage{
get_trigger(accountId, containerId, triggerId)
}
\value{
a data frame of details
}
\description{
Get a trigger's details
}
\seealso{
Other get functions: \code{\link{get_accounts_list}},
  \code{\link{get_container_versions}},
  \code{\link{get_container}},
  \code{\link{get_enviroments_list}},
  \code{\link{get_tags_list}},
  \code{\link{get_variables_list}}
}
