% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{get_triggers_list}
\alias{get_triggers_list}
\title{Get triggers list}
\usage{
get_triggers_list(accountId, containerId)
}
\value{
a list of triggers
}
\description{
Get triggers list
}
\details{

}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}},
  \code{\link{gtm_variables_list}}
}
