% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_variables_list}
\alias{get_variables_list}
\title{Get variables list}
\usage{
get_variables_list(accountId, containerId)
}
\value{
a list of variables
}
\description{
Get variables list
}
\details{

}
\seealso{
Other get functions: \code{\link{get_accounts_list}},
  \code{\link{get_container_versions}},
  \code{\link{get_container}}, \code{\link{get_tags_list}},
  \code{\link{get_triggers_list}}
}
