% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{gtm_accounts_get}
\alias{gtm_accounts_get}
\title{List account metadata}
\usage{
gtm_accounts_get(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
This returns a list containing the metadata about a single account.
If you want to get the information for all accounts, use \code{gtm_accounts_list}
}
\examples{
\dontrun{

account_id <- 12345678
account <- gtm_accounts_get(account_id)

}

}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/get}

Other account functions: 
\code{\link{gtm_accounts_list}()},
\code{\link{gtm_accounts_update}()}
}
\concept{account functions}
