% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{gtm_accounts_list}
\alias{gtm_accounts_list}
\title{See all the accounts that you have access to in GTM, view specific accounts within GTM and update account
details}
\usage{
gtm_accounts_list()
}
\description{
See all the accounts that you have access to in GTM, view specific accounts within GTM and update account
details
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts
}
\concept{account functions}
