% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm_auth.R
\name{gtm_auth}
\alias{gtm_auth}
\title{Auth for GTM Specific Properties}
\usage{
gtm_auth(token = NULL, new_user = FALSE, no_auto = FALSE)
}
\description{
Auth for GTM Specific Properties
}
\concept{login functions}
