% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_container}
\alias{gtm_container}
\title{Get a GTM account's Container detail}
\usage{
gtm_container(accountId, containerId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{GTM containerId}
}
\value{
container info
}
\description{
Get a GTM account's Container detail
}
\seealso{
Other get functions: \code{\link{get_triggers_list}},
  \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_variables_list}}
}
