% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_container_versions}
\alias{gtm_container_versions}
\title{Get a GTM account's Container version detail}
\usage{
gtm_container_versions(accountId, containerId, headers, includeDeleted)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{GTM containerId}

\item{headers}{Retrieve headers only when true. (Default: false)}

\item{includeDeleted}{Also retrieve deleted (archived) versions when true. (Default: false)}
}
\value{
container info
}
\description{
Get a GTM account's Container version detail
}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container}}, \code{\link{gtm_tag}},
  \code{\link{gtm_triggers_list}},
  \code{\link{gtm_trigger}},
  \code{\link{gtm_variables_list}}
}
