% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_containers_delete}
\alias{gtm_containers_delete}
\title{Deletes an existing GTM Container.}
\usage{
gtm_containers_delete(account_id, container_id, force = c("TRUE", "FALSE"))
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{force}{Force deletion without user input}
}
\description{
This deletes a GTM container from an account. This is an irreversible process, so it's recommended that you first back up the container.

 @export
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/deelete}

Other container functions: 
\code{\link{gtm_containers_create}()},
\code{\link{gtm_containers_get}()},
\code{\link{gtm_containers_list}()},
\code{\link{gtm_containers_update}()}
}
\concept{container functions}
