% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_containers_list}
\alias{gtm_containers_list}
\title{Manage containers in GTM
API Documentation from @seealso https://developers.google.com/tag-manager/api/v2/reference/accounts/containers}
\usage{
gtm_containers_list(account_id)
}
\description{
Manage containers in GTM
API Documentation from @seealso https://developers.google.com/tag-manager/api/v2/reference/accounts/containers
}
\concept{container functions}
