% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviroments.R
\name{gtm_enviroments_list}
\alias{gtm_enviroments_list}
\alias{gtm_enviroments_list}
\title{Get enviroments list}
\usage{
gtm_enviroments_list(accountId, containerId, environmentId)

gtm_enviroments_list(accountId, containerId, environmentId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{containerId to request}

\item{accountId}{accountId to request}

\item{containerId}{containerId to request}
}
\value{
a list of enviroments

a list of enviroments
}
\description{
Get enviroments list

Get a GTM's enviroment details
}
\details{



}
\seealso{
Other list functions: \code{\link{gtm_containers_list}},
  \code{\link{gtm_tags_list}}

Other list functions: \code{\link{gtm_containers_list}},
  \code{\link{gtm_tags_list}}
}
