% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{gtm_environments_get}
\alias{gtm_environments_get}
\title{This gets the metadata for a single GTM Environment}
\usage{
gtm_environments_get(account_id, container_id, environment_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{environment_id}{Environment Id}
}
\description{
This gets the metadata for a single GTM Environment

 @export
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/environments/get}

Other environment functions: 
\code{\link{gtm_environments_create}()},
\code{\link{gtm_environments_delete}()},
\code{\link{gtm_environments_list}()},
\code{\link{gtm_environments_reauthorize}()},
\code{\link{gtm_environments_update}()}
}
\concept{environment functions}
