% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{gtm_environments_list}
\alias{gtm_environments_list}
\title{Manage Environments within GTM}
\usage{
gtm_environments_list(account_id, container_id)
}
\description{
Manage Environments within GTM
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/environments
}
\concept{container environment functions}
