% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{gtm_environments_reauthorize}
\alias{gtm_environments_reauthorize}
\title{This reauthorises an existing GTM environment}
\usage{
gtm_environments_reauthorize(
  account_id,
  container_id,
  environment_id,
  environment
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{environment_id}{Environment Id}

\item{environment}{An Environment object}
}
\description{
@export
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/environments/reauthorize}

Other environment functions: 
\code{\link{gtm_environments_create}()},
\code{\link{gtm_environments_delete}()},
\code{\link{gtm_environments_get}()},
\code{\link{gtm_environments_list}()},
\code{\link{gtm_environments_update}()}
}
\concept{environment functions}
