% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_folders_delete}
\alias{gtm_folders_delete}
\title{Deletes a GTM Folder}
\usage{
gtm_folders_delete(
  account_id,
  container_id,
  workspace_id,
  folder_id,
  force = c("TRUE", "FALSE")
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{folder_id}{Folder Id}

\item{force}{Force deletion without user input}
}
\description{
This deletes a folder in the specified workspaces
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
workspaceId <- 3

gtm_folders_delete(accountId, containerId, workspaceId, force=TRUE)

}

}
\seealso{
\url{https://developers.google.com/tag-platform/tag-manager/api/v2/reference/accounts/containers/workspaces/folders/delete}

Other folder functions: 
\code{\link{gtm_folders_create}()},
\code{\link{gtm_folders_entities}()},
\code{\link{gtm_folders_get}()},
\code{\link{gtm_folders_list}()},
\code{\link{gtm_folders_move}()},
\code{\link{gtm_folders_revert}()},
\code{\link{gtm_folders_update}()}
}
\concept{folder functions}
