% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_folders_get}
\alias{gtm_folders_get}
\title{Gets a GTM Folder}
\usage{
gtm_folders_get(account_id, container_id, workspace_id, folder_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{folder_id}{Folder Id}
}
\description{
This returns a list containing all the metadata for a single folder in a workspace
If you want to get the information for all folders, use \code{gtm_folders_list}
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
workspaceId <- 3
folderId <- 1

folder <- gtm_folders_list(accountId, containerId, workspaceId, folderId)

}

}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/folders/get}

Other folder functions: 
\code{\link{gtm_folders_create}()},
\code{\link{gtm_folders_delete}()},
\code{\link{gtm_folders_entities}()},
\code{\link{gtm_folders_list}()},
\code{\link{gtm_folders_move}()},
\code{\link{gtm_folders_revert}()},
\code{\link{gtm_folders_update}()}
}
\concept{folder functions}
