% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_folders_list}
\alias{gtm_folders_list}
\title{Manage Folders within GTM}
\usage{
gtm_folders_list(account_id, container_id, workspace_id)
}
\description{
Manage Folders within GTM
}
\seealso{
https://developers.google.com/folder-manager/api/v2/reference/accounts/containers/workspaces/folders
}
\concept{workspace folder functions}
