% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_folders_revert}
\alias{gtm_folders_revert}
\title{Reverts a GTM Folder}
\usage{
gtm_folders_revert(account_id, container_id, workspace_id, folder_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{folder_id}{Folder Id}
}
\description{
This reverts any changes to a folder in the current workspace
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
workspaceId <- 3
folderId <- 1

folder <- gtm_folders_revert(accountId, containerId, workspaceId, folderId)

}

}
\seealso{
\url{https://developers.google.com/tag-platform/tag-manager/api/v2/reference/accounts/containers/workspaces/folders/revert}

Other folder functions: 
\code{\link{gtm_folders_create}()},
\code{\link{gtm_folders_delete}()},
\code{\link{gtm_folders_entities}()},
\code{\link{gtm_folders_get}()},
\code{\link{gtm_folders_list}()},
\code{\link{gtm_folders_move}()},
\code{\link{gtm_folders_update}()}
}
\concept{folder functions}
