% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_headers.R
\name{gtm_headers_latest}
\alias{gtm_headers_latest}
\title{Gets the latest container version header}
\usage{
gtm_headers_latest(account_id, container_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}
}
\description{
This will return the latest container version header
If you want to get the information for the most all container versions, use \code{gtm_headers_list}
}
\examples{
\dontrun{

accountId <- 1234567
containerId <- 7654321


latestHeader <- gtm_headers_latest(accountId, containerId)


}

}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/version_headers/latest}

Other version header functions: 
\code{\link{gtm_headers_list}()}
}
\concept{version header functions}
