% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_headers.R
\name{gtm_headers_list}
\alias{gtm_headers_list}
\title{Manage version headers}
\usage{
gtm_headers_list(account_id, container_id, includeDeleted = c("TRUE",
  "FALSE"))
}
\description{
Manage version headers
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/version_headers

Other container version functions: \code{\link{gtm_versions_get}}
}
\concept{container version functions}
