% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_tag}
\alias{gtm_tag}
\title{Get tags list}
\usage{
gtm_tag(accountId, containerId, tagId)
}
\value{
a data frame of tags
}
\description{
Get tags list
}
\details{

}
\seealso{
Other list functions: \code{\link{gtm_containers_list}},
  \code{\link{gtm_enviroments_list}},
  \code{\link{gtm_tags_list}}
}
