% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_tags_revert}
\alias{gtm_tags_revert}
\title{Reverts a tag,}
\usage{
gtm_tags_revert(account_id, container_id, workspace_id, tag_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{tag_id}{Tag Id}
}
\description{
This reverts a tag back to its original, unmodified state.
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/tags/revert}

Other tag structure functions: 
\code{\link{gtm_tags_create}()},
\code{\link{gtm_tags_delete}()},
\code{\link{gtm_tags_get}()},
\code{\link{gtm_tags_list}()},
\code{\link{gtm_tags_update}()}
}
\concept{tag structure functions}
