% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_tags_update}
\alias{gtm_tags_update}
\title{Update an existing tag}
\usage{
gtm_tags_update(account_id, container_id, workspace_id, tag_id, tag_object)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{tag_id}{Tag Id}

\item{tag_object}{Tag Object}
}
\description{
This takes a tag resource and overwrites the existing tag in GTM
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/tags/update}

Other tag structure functions: 
\code{\link{gtm_tags_create}()},
\code{\link{gtm_tags_delete}()},
\code{\link{gtm_tags_get}()},
\code{\link{gtm_tags_list}()},
\code{\link{gtm_tags_revert}()}
}
\concept{tag structure functions}
