% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{gtm_triggers_delete}
\alias{gtm_triggers_delete}
\title{Delete a trigger}
\usage{
gtm_triggers_delete(
  account_id,
  container_id,
  workspace_id,
  trigger_id,
  force = c(TRUE, FALSE)
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{trigger_id}{trigger Id}

\item{force}{Force deletion without user input}
}
\description{
This deletes a trigger in a GTM workspace
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/triggers/delete}

Other trigger structure functions: 
\code{\link{gtm_triggers_create}()},
\code{\link{gtm_triggers_get}()},
\code{\link{gtm_triggers_list}()},
\code{\link{gtm_triggers_revert}()},
\code{\link{gtm_triggers_update}()}
}
\concept{trigger structure functions}
