% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_variables_list}
\alias{gtm_variables_list}
\title{Manage Variables in GTM}
\usage{
gtm_variables_list(account_id, container_id, workspace_id)
}
\description{
Manage Variables in GTM
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/variables

Other account structure functions: \code{\link{gtm_builtin_list}}
}
\concept{account structure functions}
