% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_variables_update}
\alias{gtm_variables_update}
\title{Update an existing variable}
\usage{
gtm_variables_update(
  account_id,
  container_id,
  workspace_id,
  variable_id,
  variable_object
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{variable_id}{Variable Id}

\item{variable_object}{Variable Object}
}
\description{
This takes a variable resource and overwrites the existing variable in GTM
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/variables/update}

Other variable structure functions: 
\code{\link{gtm_variables_create}()},
\code{\link{gtm_variables_delete}()},
\code{\link{gtm_variables_get}()},
\code{\link{gtm_variables_list}()},
\code{\link{gtm_variables_revert}()}
}
\concept{variable structure functions}
