% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_get}
\alias{gtm_versions_get}
\title{Manage Versions in GTM}
\usage{
gtm_versions_get(account_id, container_id, version_id)
}
\description{
Manage Versions in GTM
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/versions

Other container version functions: \code{\link{gtm_headers_list}}
}
\concept{container version functions}
