% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_publish}
\alias{gtm_versions_publish}
\title{Publish a new container version}
\usage{
gtm_versions_publish(account_id, container_id, version_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{version_id}{Version Id}
}
\description{
Publishes a container version to be live
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
versionId <- 6

version <- gtm_versions_publish(accountId, containerId, versionId)

}


}
\seealso{
\url{https://developers.google.com/tag-platform/tag-manager/api/v2/reference/accounts/containers/versions/publish}

Other versions functions: 
\code{\link{gtm_versions_setlatest}()},
\code{\link{gtm_versions_undelete}()}
}
\concept{versions functions}
