% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_update}
\alias{gtm_versions_update}
\title{Update a container version}
\usage{
gtm_versions_update(
  account_id,
  container_id,
  version_id,
  name = NULL,
  description = NULL
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{version_id}{Version Id}

\item{name}{Version Name}

\item{description}{Version Description}
}
\description{
Updates a container Version
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
versionId <- 6
newName <- 'My Cool New Version'
newDescription <- 'My new version does something cool'
version <- gtm_versions_update(accountId, containerId, versionId, newName, newDescription)

}

}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/versions/update}

Other version functions: 
\code{\link{gtm_versions_delete}()},
\code{\link{gtm_versions_get}()}
}
\concept{version functions}
