% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{gtm_workspaces_list}
\alias{gtm_workspaces_list}
\title{Manage Workspaces in GTM}
\usage{
gtm_workspaces_list(account_id, container_id)
}
\description{
Manage Workspaces in GTM
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces
}
\concept{workspaces functions}
