% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_builtin.R
\name{variables_list}
\alias{variables_list}
\title{Built-In Variable Keys}
\arguments{
\item{advertiserId}{Advertiser Id - Firebase}

\item{advertisingTrackingEnabled}{Advertiser Tracking Enabled - Firebase}

\item{ampBrowserLanguage}{Browser Language - AMP}

\item{ampCanonicalHost}{Cannonical Host - AMP}

\item{ampCanonicalPath}{Canonical Path - AMP}

\item{ampCanonicalUrl}{Canonical URL - AMP}

\item{ampClientId}{Page Client ID - AMP}

\item{ampClientMaxScrollX}{Scroll Width - AMP}

\item{ampClientMaxScrollY}{Scroll Height - AMP}

\item{ampClientScreenHeight}{Screen Height - AMP}

\item{ampClientScreenWidth}{Screen Width - AMP}

\item{ampClientScrollX}{Scroll Left - AMP}

\item{ampClientScrollY}{Scroll Top - AMP}

\item{ampClientTimestamp}{Client Timestamp - AMP}

\item{ampClientTimezone}{Client Timezone - AMP}

\item{ampGtmEvent}{AMP Event - AMP}

\item{ampPageDownloadTime}{Page Downlad Time - AMP}

\item{ampPageLoadTime}{Page Load Time - AMP}

\item{ampPageViewId}{Page View Id  - AMP}

\item{ampReferrer}{Document Referrer - AMP}

\item{ampTitle}{Document Title - AMP}

\item{ampTotalEngagedTime}{Total Engaged Time - AMP}

\item{appId}{App Id - Firebase}

\item{appName}{App Name- Firebase}

\item{appVersionCode}{App Version Code - Firebase}

\item{appVersionName}{App Version Name - Firebase}

\item{builtInVariableTypeUnspecified}{Undefined}

\item{clickClasses}{Click Classes}

\item{clickElement}{Click Element}

\item{clickId}{Click Id}

\item{clickTarget}{Click Target}

\item{clickText}{Click Text}

\item{clickUrl}{Click URL}

\item{containerId}{Container ID}

\item{containerVersion}{Container Version}

\item{debugMode}{Debug Mode}

\item{deviceName}{Device Name}

\item{elementVisibilityFirstTime}{Element First Visible}

\item{elementVisibilityRatio}{Percent Visible}

\item{elementVisibilityRecentTime}{Element Last Visible}

\item{elementVisibilityTime}{On Screen Duration}

\item{environmentName}{Environment Name}

\item{errorLine}{Error Line}

\item{errorMessage}{Error Message}

\item{errorUrl}{Error URL}

\item{event}{Event}

\item{eventName}{Event Name}

\item{firebaseEventParameterCampaign}{Campaign  Name - Firebase}

\item{firebaseEventParameterCampaignAclid}{Click Id (aclid) - Firebase}

\item{firebaseEventParameterCampaignAnid}{Ad Network ID (anid) - Firebase}

\item{firebaseEventParameterCampaignClickTimestamp}{Campaign Click Timestamp - Firebase}

\item{firebaseEventParameterCampaignContent}{Campaign Content - Firebase}

\item{firebaseEventParameterCampaignCp1}{Campaign Custom Parameter 1 - Firebase}

\item{firebaseEventParameterCampaignGclid}{Google Click Id (gclid) - Firebase}

\item{firebaseEventParameterCampaignSource}{Campaign Source - Firebase}

\item{firebaseEventParameterCampaignTerm}{Campaign Search Term  - Firebase}

\item{firebaseEventParameterCurrency}{In-app Purchase Currency - Firebase}

\item{firebaseEventParameterDynamicLinkAcceptTime}{Dynamic Link Accept Time - Firebase}

\item{firebaseEventParameterDynamicLinkLinkid}{Dynamic Link ID - Firebase}

\item{firebaseEventParameterNotificationMessageDeviceTime}{Notification Device Time - Firebase}

\item{firebaseEventParameterNotificationMessageId}{Notification Message Id - Firebase}

\item{firebaseEventParameterNotificationMessageName}{Notification Message Name - Firebase}

\item{firebaseEventParameterNotificationMessageTime}{Notification Message Time - Firebase}

\item{firebaseEventParameterNotificationTopic}{Notification Message Topic - Firebase}

\item{firebaseEventParameterPreviousAppVersion}{Previous Application Version - Firebase}

\item{firebaseEventParameterPreviousOsVersion}{Previous OS Version - Firebase}

\item{firebaseEventParameterPrice}{In-app Purchase Price - Firebase}

\item{firebaseEventParameterProductId}{In-app Purchase Product ID - Firebase}

\item{firebaseEventParameterQuantity}{In-app Purchase Quantity - Firebase}

\item{firebaseEventParameterValue}{In-app Purchase Value - Firebase}

\item{formClasses}{Form Classes}

\item{formElement}{Form Element}

\item{formId}{Form Id}

\item{formTarget}{Form Target}

\item{formText}{Form Text}

\item{formUrl}{Form URL}

\item{historySource}{History Source}

\item{htmlId}{HTML ID}

\item{language}{Language}

\item{newHistoryFragment}{New History Fragment}

\item{newHistoryState}{New History State}

\item{newHistoryUrl}{New History URL}

\item{oldHistoryFragment}{Old History Fragment}

\item{oldHistoryState}{Old History Stat}

\item{oldHistoryUrl}{Old History URL}

\item{osVersion}{OS Version}

\item{pageHostname}{Page Hostname}

\item{pagePath}{Page Path}

\item{pageUrl}{Page URL}

\item{platform}{Platform}

\item{randomNumber}{Random Number}

\item{referrer}{Referrer}

\item{resolution}{Screen Resolution}

\item{scrollDepthDirection}{Scroll Depth Direction}

\item{scrollDepthThreshold}{Scroll Depth Threshold}

\item{scrollDepthUnits}{Scroll Depth Units}

\item{sdkVersion}{SDK Version - Firebase}

\item{videoCurrentTime}{Video Current Time}

\item{videoDuration}{Video Duration}

\item{videoPercent}{Video Percent}

\item{videoProvider}{Video Provider}

\item{videoStatus}{Video Status}

\item{videoTitle}{Video Title}

\item{videoUrl}{Video URL}

\item{videoVisible}{Video Visible}
}
\description{
These are the variable names accepted when enabling/disabling Built-In Variables
}
\seealso{
Other built in variable functions: 
\code{\link{gtm_builtin_create}()},
\code{\link{gtm_builtin_delete}()},
\code{\link{gtm_builtin_list}()},
\code{\link{gtm_builtin_revert}()}
}
\concept{built in variable functions}
