% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_accounts_list}
\alias{get_accounts_list}
\title{Get account list}
\usage{
get_accounts_list()
}
\value{
a dataframe of accounts
}
\description{
Get account list
}
\details{

}
\seealso{
Other managementAPI functions: \code{\link{get_containers_list}},
  \code{\link{get_tags_list}},
  \code{\link{get_triggers_list}},
  \code{\link{get_variables_list}}
}
