% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviroments.R
\name{get_enviroments_list}
\alias{get_enviroments_list}
\alias{get_enviroments_list}
\title{Get enviroments list}
\usage{
get_enviroments_list(accountId, containerId, environmentId)

get_enviroments_list(accountId, containerId, environmentId)
}
\value{
a list of enviroments

a list of enviroments
}
\description{
Get enviroments list

Get a GTM's enviroment details
}
\details{



}
\seealso{
Other list functions: \code{\link{get_containers_list}}

Other list functions: \code{\link{get_containers_list}}
}
