% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags_list}
\alias{get_tags_list}
\title{Get tags list}
\usage{
get_tags_list(accountId, containerId)
}
\value{
a data frame of tags
}
\description{
Get tags list
}
\details{

}
\seealso{
Other get functions: \code{\link{get_accounts_list}},
  \code{\link{get_container_versions}},
  \code{\link{get_container}},
  \code{\link{get_triggers_list}},
  \code{\link{get_variables_list}}
}
