% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{get_triggers_list}
\alias{get_triggers_list}
\title{Get triggers list}
\usage{
get_triggers_list(accountId, containerId)
}
\value{
a list of triggers
}
\description{
Get triggers list
}
\seealso{
Other list functions: \code{\link{get_containers_list}},
  \code{\link{get_enviroments_list}}
}
