% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{gtm_accounts_list}
\alias{gtm_accounts_list}
\title{Get GTM account list}
\usage{
gtm_accounts_list()
}
\value{
a dataframe of accounts
}
\description{
Get GTM account list
}
\details{

}
\seealso{
Other get functions: \code{\link{get_triggers_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}},
  \code{\link{gtm_variables_list}}
}
