% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{gtm_accounts_list}
\alias{gtm_accounts_list}
\title{List all accounts a user has accesss to}
\usage{
gtm_accounts_list()
}
\description{
This will return a data frame all your available accounts, returning the path, account Id and account name.
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/list}

Other account functions: 
\code{\link{gtm_accounts_get}()},
\code{\link{gtm_accounts_update}()}
}
\concept{account functions}
