% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtm_auth.R
\name{gtm_auth}
\alias{gtm_auth}
\title{Auth for GTM Specific Properties}
\usage{
gtm_auth(token = NULL)
}
\arguments{
\item{token}{An existing Google Auth Token}
}
\description{
This function authenticates the user with Google Tag Manager
}
\concept{login functions}
