% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_containers_list}
\alias{gtm_containers_list}
\title{Get containers list}
\usage{
gtm_containers_list(accountId)
}
\arguments{
\item{accountId}{accountId to request}
}
\value{
a list of containers
}
\description{
Get containers list
}
\seealso{
Other list functions: \code{\link{gtm_enviroments_list}},
  \code{\link{gtm_tags_list}}
}
