% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{gtm_environments_list}
\alias{gtm_environments_list}
\title{List the containers}
\usage{
gtm_environments_list(account_id, container_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}
}
\description{
This lists the environments in a specified container.

 @export
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/environments/list}

Other environment functions: 
\code{\link{gtm_environments_create}()},
\code{\link{gtm_environments_delete}()},
\code{\link{gtm_environments_get}()},
\code{\link{gtm_environments_reauthorize}()},
\code{\link{gtm_environments_update}()}
}
\concept{environment functions}
