% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_tag}
\alias{gtm_tag}
\title{Get tags list}
\usage{
gtm_tag(accountId, containerId, tagId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{containerId to request}

\item{tagId}{tagId to request}
}
\value{
a data frame of tags
}
\description{
Get tags list
}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}},
  \code{\link{gtm_triggers_list}},
  \code{\link{gtm_trigger}},
  \code{\link{gtm_variables_list}}
}
