% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_tags_list}
\alias{gtm_tags_list}
\title{Get tags list}
\usage{
gtm_tags_list(accountId, containerId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{containerId to request}
}
\value{
a data frame of tags
}
\description{
Get tags list
}
\details{

}
\seealso{
Other list functions: \code{\link{gtm_containers_list}},
  \code{\link{gtm_enviroments_list}}
}
