% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{gtm_trigger}
\alias{gtm_trigger}
\title{Get a triggers details}
\usage{
gtm_trigger(accountId, containerId, triggerId)
}
\value{
a daa frame of trigger detials
}
\description{
Get a triggers details
}
\details{

}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}}, \code{\link{gtm_tag}},
  \code{\link{gtm_triggers_list}},
  \code{\link{gtm_variables_list}}
}
