% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{gtm_triggers_list}
\alias{gtm_triggers_list}
\title{Get triggers list}
\usage{
gtm_triggers_list(accountId, containerId)
}
\value{
a list of triggers
}
\description{
Get triggers list
}
\details{

}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}}, \code{\link{gtm_tag}},
  \code{\link{gtm_trigger}},
  \code{\link{gtm_variables_list}}
}
