% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_variables_list}
\alias{gtm_variables_list}
\title{Get variables list}
\usage{
gtm_variables_list(accountId, containerId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{containerId to request}
}
\value{
a list of variables
}
\description{
Get variables list
}
\details{

}
\seealso{
Other get functions: \code{\link{gtm_accounts_list}},
  \code{\link{gtm_container_versions}},
  \code{\link{gtm_container}}, \code{\link{gtm_tag}},
  \code{\link{gtm_triggers_list}},
  \code{\link{gtm_trigger}}
}
