% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_delete}
\alias{gtm_versions_delete}
\title{Delete a container version}
\usage{
gtm_versions_delete(
  account_id,
  container_id,
  version_id,
  force = c("TRUE", "FALSE")
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{version_id}{Version Id}

\item{force}{Force deletion without user input}
}
\description{
Deletes a container version
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/versions/delete}

Other version structure functions: 
\code{\link{gtm_versions_get}()},
\code{\link{gtm_versions_update}()}
}
\concept{version structure functions}
