% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_get}
\alias{gtm_versions_get}
\title{List a single version in a workspace}
\usage{
gtm_versions_get(account_id, container_id, version_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{version_id}{Version Id}
}
\description{
This returns a single
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/versions/get}

Other version structure functions: 
\code{\link{gtm_versions_delete}()},
\code{\link{gtm_versions_update}()}
}
\concept{version structure functions}
