% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_versions_setlatest}
\alias{gtm_versions_setlatest}
\title{Sets a container version as the latest version}
\usage{
gtm_versions_setlatest(account_id, container_id, version_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{version_id}{Version Id}
}
\description{
Sets the given container version as the latest version
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
versionId <- 6

version <- gtm_versions_setlatest(accountId, containerId, versionId)

}


}
\seealso{
\url{https://developers.google.com/tag-platform/tag-manager/api/v2/reference/accounts/containers/versions/set_latest}

Other versions functions: 
\code{\link{gtm_versions_publish}()},
\code{\link{gtm_versions_undelete}()}
}
\concept{versions functions}
