% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{gtm_workspaces_resolve}
\alias{gtm_workspaces_resolve}
\title{Resolve workspace conflicts}
\usage{
gtm_workspaces_resolve(
  account_id,
  container_id,
  workspace_id,
  tag = NULL,
  trigger = NULL,
  variable = NULL,
  folder = NULL,
  changeStatus = c("added", "changeStatusUnspecified", "deleted", "none", "updated")
)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}

\item{tag}{A Tag Object}

\item{trigger}{A Trigger Object}

\item{variable}{A Variable Object}

\item{folder}{A Folder Object}

\item{changeStatus}{Represents how the entity has been changed in the workspace.}
}
\description{
Resolves a merge conflict for a workspace entity by updating it to the resolved 
entity passed in the request.
}
\seealso{
\url{https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/resolve_conflict}

Other workspace functions: 
\code{\link{gtm_workspaces_create}()},
\code{\link{gtm_workspaces_delete}()},
\code{\link{gtm_workspaces_export}()},
\code{\link{gtm_workspaces_get}()},
\code{\link{gtm_workspaces_list}()},
\code{\link{gtm_workspaces_preview}()},
\code{\link{gtm_workspaces_status}()},
\code{\link{gtm_workspaces_submit}()},
\code{\link{gtm_workspaces_sync}()},
\code{\link{gtm_workspaces_update}()}
}
\concept{workspace functions}
