% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{gtm_workspaces_sync}
\alias{gtm_workspaces_sync}
\title{Bring a workspace in line with the latest version}
\usage{
gtm_workspaces_sync(account_id, container_id, workspace_id)
}
\arguments{
\item{account_id}{Account Id}

\item{container_id}{Container Id}

\item{workspace_id}{Workspace Id}
}
\description{
Syncs a workspace to the latest container version by updating all unmodified 
workspace entities and displaying conflicts for modified entities
}
\examples{

\dontrun{
accountId <- 1234567
containerId <- 7654321
workspaceId <- 6

syncResult <- gtm_workspaces_sync(accountId, containerId, workspaceId)

}


}
\seealso{
\url{https://developers.google.com/tag-platform/tag-manager/api/v2/reference/accounts/containers/workspaces/sync}

Other workspace functions: 
\code{\link{gtm_workspaces_create}()},
\code{\link{gtm_workspaces_delete}()},
\code{\link{gtm_workspaces_get}()},
\code{\link{gtm_workspaces_list}()},
\code{\link{gtm_workspaces_preview}()},
\code{\link{gtm_workspaces_resolve}()},
\code{\link{gtm_workspaces_status}()},
\code{\link{gtm_workspaces_submit}()},
\code{\link{gtm_workspaces_update}()}
}
\concept{workspace functions}
