% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gtm_zones_list}
\alias{gtm_zones_list}
\title{Manage Zones in GTM}
\usage{
gtm_zones_list(account_id, container_id, workspace_id)
}
\description{
Manage Zones in GTM
}
\seealso{
https://developers.google.com/tag-manager/api/v2/reference/accounts/containers/workspaces/zones
}
\concept{zone functions}
